Refactor: update invoice and booking models with type changes and imports cleanup
