+++ easyverein/models/booking.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 from pydantic import BaseModel
 
 from ..core.types import DateTime, EasyVereinReference, FilterIntList, Sphere
-from . import BookingProject
 from .base import EasyVereinBase
 from .mixins.empty_strings_mixin import EmptyStringsToNone
 from .mixins.required_attributes import required_mixin
@@ -101,4 +100,5 @@ class BookingFilter(BaseModel):
 
 
 from .billing_account import BillingAccount  # noqa: E402
+from .booking_project import BookingProject  # noqa: E402
 from .invoice import Invoice  # noqa: E402
